% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_qli_df.R
\name{construct_qli_df}
\alias{construct_qli_df}
\title{A helper function for constructing a dataset with QLI values from dataframes
containing lower and upper quantile estimates}
\usage{
construct_qli_df(q_lower, q_upper, predictor)
}
\arguments{
\item{q_lower}{A dataset containing output from quantile smoother for lower quantiles}

\item{q_upper}{A dataset containing output from quantile smoother for upper quantiles}

\item{predictor}{A character. Name of the predictor as it appears in the dataset used for fitting \code{mod}. If predictor is "fitted" (default), fitted values from the \code{mod} are used instead.}
}
\value{
A dataframe with QLI values which can be used to fit a model predicting the change in the width of the interval.
}
\description{
A helper function for constructing a dataset with QLI values from dataframes
containing lower and upper quantile estimates
}
\examples{
\dontrun{
construct_qli_df(q_lower, q_upper, "predictor_name")
}
}
