% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creativity_data.R
\docType{data}
\name{creativity_data}
\alias{creativity_data}
\title{Creativity data from Meier et al (2020)}
\format{
A tibble with 59 rows and 3 variables:
\describe{
\item{divergent_thinking}{Numeric. Originally labelled Fluency.}
\item{relaxation}{Numeric. Originally labelled AUCi_RSA_36}
\item{creativity}{Numeric. Originally labelled CAQ_score.}
}
}
\source{
\url{https://econtent.hogrefe.com/doi/abs/10.1027/1618-3169/a000483?journalCode=zea}
}
\usage{
creativity_data
}
\description{
The original dataset can be located at: https://osf.io/d4s6q/
}
\details{
The version of the data was processed to only include variables necessary for reproducing the linear model used for demonstration in https://doi.org/10.31234/osf.io/gn4mr .
}
\keyword{datasets}
