% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qli.R
\name{qli}
\alias{qli}
\title{Quantify heteroscedasticity in linear models using Quantile LOWESS Intervals (QLI)}
\usage{
qli(
  mod,
  predictor = "fitted",
  standardise = TRUE,
  plot = FALSE,
  plot_args = list(interval_fill_col = "#1887ac", interval_outline_col = "#1887ac",
    point_col = "black"),
  lower_quant = 0.025,
  upper_quant = 0.975,
  window_prop = 0.1,
  prop_overlap = 0.75
)
}
\arguments{
\item{mod}{A linear model object fitted with lm()}

\item{predictor}{A character vector. Name of the predictor as it appears in the dataset used for fitting \code{mod}. If predictor is "fitted" (default), fitted values from the \code{mod} are used instead.}

\item{standardise}{If TRUE (default) predictor values and model residuals will be standardised.}

\item{plot}{If TRUE, output includes a scatterplot of predictor vs residuals with the Quantile LOWESS Interval. Default is FALSE.}

\item{plot_args}{List of arguments to customise the plot. Valid arguments include, \code{interval_fill_col} for the fill colour of the interval, \code{interval_outline_col} for the colour of the interval outline, and \code{point_col} for the colour of the points on the scatter plot. Further customisation can be achieved by adding \code{ggplot2} layers. Alternatively, run function \code{qli_values()} instead to produce a dataset with values for plotting.}

\item{lower_quant}{A quantile to use for the lower boundary of the quantile lowess interval. A numeric value between 0 and 1. Default is 0.025 (recommended).}

\item{upper_quant}{A quantile to use for the upper boundary of the quantile lowess interval. A numeric value between 0 and 1. Default is 0.975 (recommended).}

\item{window_prop}{Proportion of the total sample size use in each rolling window. A numeric value between 0 and 1. Default is 0.10 (recommended).}

\item{prop_overlap}{The extent to which rolling windows can overlap. A numeric value between 0 and 1. Default is 0.75 (recommended).}
}
\value{
If plot is TRUE, returns a list with polynomial trends estimating changes in QLI width and a plot. If plot is false, returns only QLI results.
}
\description{
Quantify heteroscedasticity in linear models using Quantile LOWESS Intervals (QLI)
}
\examples{
cars_mod <- lm(dist ~ speed, data = cars)
qli(mod = cars_mod, predictor = "speed")
}
