% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_smoother.R
\name{quantile_smoother}
\alias{quantile_smoother}
\title{Calculate quantile values of y across the values of x}
\usage{
quantile_smoother(
  y,
  x,
  prop_overlap = 0.75,
  window_prop = 0.05,
  tau = 0.95,
  window_alignment = c("center"),
  window_function = function(x) {
     stats::quantile(x, tau)
 }
)
}
\arguments{
\item{y}{A numeric vector representing the outcome}

\item{x}{A numeric vector representing the predictor}

\item{prop_overlap}{The extent to which rolling windows can overlap. A numeric value between 0 and 1.}

\item{window_prop}{Proportion of the total sample size use in each rolling window. A numeric value between 0 and 1.}

\item{tau}{Quantile through which to fit the LOWESS smoother. A numeric value between 0 and 1.}

\item{window_alignment}{Alignment for rolling windows. The default is "center". Other options are "left" and "right".}

\item{window_function}{Function to apply in each rolling window. The default is \code{function(x) {stats::quantile(x, tau)}} which obtains the quantile values.}
}
\value{
A list with two elements. x representing the predictor values and y.lowess representing the smoothed quantile LOWESS values
}
\description{
Calculate quantile values of y across the values of x
}
\examples{
quantile_smoother(x = cars$speed, y = cars$dist)
}
