% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z.R
\name{z}
\alias{z}
\title{Convert a vector into Z scores}
\usage{
z(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A numeric vector
}
\description{
Convert a vector into Z scores
}
\examples{
x <- c(1,2,3,4,5)
z(x)
}
